Provide image_path that works with Sprockets
Prior to this, functions in sprockets-sass conflicted with ours and caused an infinite loop and StackError. This fixes thomas-mcdonald/bootstrap-sass#257

Note that Sprockets' asset_path requires configuration if it's not already configured by your framework. For example, in a simple Rack app, add the following to your config.ru:

environment.context_class.class_eval do
  def asset_path(path, options = {})
    "/assets/#{path}"
  end
end